+++ src/prefect/configuration.py
@@ -1,3 +1,4 @@
+import inspect
 import datetime
 import logging
 import os
@@ -245,9 +246,36 @@ def process_task_defaults(config: Config) -> Config:
 
 def validate_config(config: Config) -> None:
     """
-    Placeholder for future config validation. For example, invalid values could raise an error.
+    Validates that the configuration file is valid.
+        - keys are lowercase
+        - keys do not shadow Config methods
+
+    Note that this is performed when the config is first loaded, but not after.
     """
-    pass
+
+    def check_lowercase_keys(config: Config) -> None:
+        """
+        Recursively check that keys are lowercase
+        """
+        for k, v in config.items():
+            if k != k.lower():
+                raise ValueError('Config keys must be lowercase: "{}"'.format(k))
+            if isinstance(v, Config) and k != "context":
+                check_lowercase_keys(v)
+
+    def check_valid_keys(config: Config) -> None:
+        """
+        Recursively check that keys do not shadow methods of the Config object
+        """
+        invalid_keys = dir(Config)
+        for k, v in config.items():
+            if k in invalid_keys:
+                raise ValueError('Invalid config key: "{}"'.format(k))
+            if isinstance(v, Config):
+                check_valid_keys(v)
+
+    check_lowercase_keys(config)
+    check_valid_keys(config)
 
 
 # Load configuration ----------------------------------------------------------
@@ -258,7 +286,7 @@ def load_toml(path: str) -> dict:
     Loads a config dictionary from TOML
     """
     return {
-        key.lower(): value
+        key: value
         for key, value in toml.load(cast(str, interpolate_env_vars(path))).items()
     }