+++ src/prefect/engine/cache_validators.py
@@ -88,7 +88,7 @@ def all_inputs(
     """
     if duration_only(state, inputs, parameters) is False:
         return False
-    elif state.cached_inputs == inputs:
+    elif {key: res.value for key, res in (state.cached_inputs or {}).items()} == inputs:
         return True
     else:
         return False
@@ -263,7 +263,9 @@ def partial_inputs_only(validate_on: Iterable[str] = None,) -> Callable:
                 True
             )  # if you dont want to validate on anything, then the cache is valid
         else:
-            cached = state.cached_inputs or {}
+            cached = {
+                key: res.value for key, res in (state.cached_inputs or {}).items()
+            }
             partial_provided = {
                 key: value for key, value in inputs.items() if key in validate_on
             }