+++ src/prefect/serialization/task.py
@@ -3,14 +3,7 @@ from collections import OrderedDict
 from typing import TYPE_CHECKING, Any, Callable, Dict
 
 import marshmallow
-from marshmallow import (
-    ValidationError,
-    fields,
-    post_dump,
-    post_load,
-    pre_dump,
-    pre_load,
-)
+from marshmallow import ValidationError, fields, post_load
 
 import prefect
 from prefect.utilities.serialization import (
@@ -55,7 +48,7 @@ class TaskMethodsMixin:
         return str(task.outputs())
 
     @post_load
-    def create_object(self, data: dict) -> prefect.core.Task:
+    def create_object(self, data: dict, **kwargs: Any) -> prefect.core.Task:
         """
         Sometimes we deserialize tasks and edges simultaneously (for example, when a
         Flow is being deserialized), in which case we check slugs to see if we already
@@ -63,10 +56,12 @@ class TaskMethodsMixin:
         cache.
         """
         slug = data.get("slug")
+        auto_generated = data.pop("auto_generated", False)
 
         # if the slug is not in the task cache, create a task object and add it
         if slug not in self.context.setdefault("task_cache", {}):  # type: ignore
             task = super().create_object(data)  # type: ignore
+            task.auto_generated = auto_generated  # type: ignore
             self.context["task_cache"][slug] = task  # type: ignore
 
         # return the task object from the cache
@@ -119,6 +114,7 @@ class TaskSchema(TaskMethodsMixin, ObjectSchema):
         reject_invalid=False,
         allow_none=True,
     )
+    auto_generated = fields.Boolean(allow_none=True)
 
 
 class ParameterSchema(TaskMethodsMixin, ObjectSchema):